requirements.txt
tensorflow>=2.10.0
python-telegram-bot>=13.0
Pillow>=9.0.0
requests>=2.28.0
numpy>=1.21.0
flask>=2.2.0 